cdc_priority/classifier/model.py
import torch
import torch.nn as nn

from thesios_classifier.model import ThesiosClassifier, ThesiosClassifierV2


class MLPClassifier(nn.Module):
    def __init__(self, input_dim: int, hidden_dim: int, num_classes: int, dropout: float):
        super().__init__()
        self.network = nn.Sequential(
            nn.Linear(input_dim, hidden_dim),
            nn.ReLU(),
            nn.Dropout(dropout),
            nn.Linear(hidden_dim, hidden_dim),
            nn.ReLU(),
            nn.Dropout(dropout),
            nn.Linear(hidden_dim, num_classes),
        )

    def forward(self, inputs: torch.Tensor) -> torch.Tensor:
        return self.network(inputs)


__all__ = [
    "MLPClassifier",
    "ThesiosClassifier",
    "ThesiosClassifierV2",
]
