requirements.txt
PyYAML>=6.0
pandas>=2.0
scikit-learn>=1.4
torch>=2.2
